+++ src/main/java/com/example/loginlivesession2/dto/response/PostResDto.java
@@ -14,7 +14,7 @@ public class PostResDto {
     private LocalDateTime createAt;
     private String nickname;
     private Long postid;
-    private int like;
+    private Long like;
     private String title;
     private String content;
     private List<CommentResDto> commentResDtoList;
